+++ examples/complex/complex/commands/cmd_status.py
@@ -1,9 +1,9 @@
 import click
-from complex.cli import pass_context
+from complex.cli import pass_environment
 
 
 @click.command('status', short_help='Shows file changes.')
-@pass_context
+@pass_environment
 def cli(ctx):
     """Shows file changes in the current working directory."""
     ctx.log('Changed files: none')